feat: add handling for projects

+++ backend/internal/utils/helper.go
@@ -9,11 +9,11 @@ import (
 
 type Envelope map[string]interface{}
 
-func ReadIdParam(r *http.Request) (int64, error) {
+func ReadIdParam(r *http.Request) (uint, error) {
 	param := mux.Vars(r)["id"]
 	id, err := strconv.Atoi(param)
 	if err != nil || id < 1 {
 		return 0, errors.New("invalid id parameter")
 	}
-	return int64(id), nil
+	return uint(id), nil
 }
